Revert "correcion metodo cupon"
This reverts commit 7a86fe405c7fd459648bdbf12d7bbf90743f42e4.

+++ src/main/java/negocio/BOs/PedidoProgramadoBO.java
@@ -67,7 +67,7 @@ public void registrarPedido(PedidoProgramadoDTO dto)
     descuentoCalculado = subtotalCalculado *
             cuponValido.getDescuento().doubleValue();
 
-    idCuponAplicado = cuponValido.getIdCupon();
+    idCuponAplicado = cuponValido.getIdCupon().intValue();
     }
 
     if (Math.abs(subtotalCalculado - dto.getSubtotal()) > 0.01) {
